Move icons : `RefundIcon`,   `GiftIcon`,  `LockOutlineIcon`
