+++ assets/javascripts/kitten/components/graphics/icons-next/cube-icon-next/index.js
@@ -4,14 +4,17 @@ import PropTypes from 'prop-types'
 export const CubeIconNext = ({ color, title, ...props }) => (
   <svg
     width="24"
-    height="34"
-    viewBox="0 0 24 34"
+    height="33"
+    viewBox="0 0 24 33"
     xmlns="http://www.w3.org/2000/svg"
     fill="none"
     {...props}
   >
     {title && <title>{title}</title>}
-    <path fill={color} stroke={color} stroke-width=".2" d="M.58 13.842h0l11.065-5.098a.907.907 0 01.71 0l11.06 5.096a.914.914 0 01.485.808v12.71c0 .324-.195.645-.48.771h0l-11.065 5.098a.907.907 0 01-.71 0L.582 28.095H.579C.296 27.966.1 27.645.1 27.322v-12.71c0-.324.196-.645.48-.77zm10.084 16.372l.02.01h.158V20.58l-.057-.027-8.192-3.806-.142-.066v9.699l.058.027 8.155 3.806zm1.294-11.705l.042.02.042-.02 7.689-3.59.195-.091-.195-.09-7.689-3.555-.042-.02-.042.02-7.689 3.59-.195.091.195.09 7.689 3.555zm1.235 11.614v.157l.143-.066 8.135-3.796h.114v-9.736l-.142.066-8.192 3.806-.058.027v9.542zM11.936 6.82h-.013c-.659-.008-1.18-.536-1.18-1.157v-3.77c0-.663.53-1.193 1.193-1.193s1.193.53 1.193 1.193v3.734c0 .662-.53 1.192-1.193 1.192zM16.878 8.236l-.035-.023c-.54-.373-.687-1.082-.334-1.596h0l.003-.004 1.976-3.195h0l.001-.002c.33-.56 1.055-.727 1.62-.362h.001c.526.329.696 1.057.363 1.623l-1.975 3.158s0 0 0 0a1.164 1.164 0 01-.993.564c-.241 0-.434-.034-.627-.163zM7.36 6.613h0l.001.002c.365.563.199 1.287-.359 1.617-.205.102-.44.167-.635.167a1.164 1.164 0 01-.994-.564L3.398 4.676h0l-.001-.001c-.365-.564-.198-1.29.362-1.619h0l.004-.002c.564-.365 1.29-.198 1.62.362h0v.002L7.36 6.613z"/>
+    <path fill={color} d="M23.461 13.15L12.395 8.053a1.007 1.007 0 00-.79 0L.539 13.151c-.323.143-.539.503-.539.862v12.71c0 .36.216.718.539.862l11.066 5.134c.251.108.539.108.79 0l11.066-5.099c.324-.143.539-.502.539-.861v-12.71c0-.396-.215-.718-.539-.898zM10.707 29.524L2.55 25.718v-9.48l8.191 3.807v9.478h-.035zM12 17.82l-7.689-3.555L12 10.674l7.689 3.554L12 17.818zm9.45 7.899l-8.157 3.805v-9.479l8.192-3.805v9.479h-.036z"/>
+    {/* <path fill="#F6F6F6" fill-rule="evenodd" d="M23.461 27.62c.324-.143.539-.502.539-.861v-12.71c0-.396-.215-.718-.539-.898L12.395 8.052a1.007 1.007 0 00-.79 0L.539 13.151c-.323.143-.539.503-.539.862v12.71c0 .36.216.718.539.862l11.066 5.134c.251.108.539.108.79 0l11.066-5.099zm-11.147 4.916l11.066-5.098c.245-.109.42-.392.42-.68v-12.71a.813.813 0 00-.43-.719L12.314 8.235a.808.808 0 00-.628 0L.62 13.334c-.244.108-.42.391-.42.679v12.71c0 .287.175.57.42.679l.003.001 11.063 5.133c.2.085.428.085.628 0zm-1.652-2.813l-8.311-3.878v-9.92l8.591 3.992v9.806h-.28zM12 18.04l-8.163-3.774L12 10.453l8.163 3.774L12 18.04zm1.093 11.798v-9.92l8.592-3.992v9.992h-.191l-8.4 3.92zm8.356-4.12h.036V16.24l-8.192 3.805v9.48l8.156-3.806zm-10.707-5.672l-8.191-3.806v9.479l8.156 3.805h.035v-9.478zM12 17.818l7.689-3.59L12 10.673l-7.689 3.59L12 17.82z" clip-rule="evenodd"/> */}
+    <path fill={color} d="M11.936 6.32c.719 0 1.293-.575 1.293-1.293V1.293C13.23.575 12.655 0 11.936 0s-1.293.575-1.293 1.293v3.77c0 .682.574 1.256 1.293 1.256zM16.822 7.72c.216.143.431.18.683.18.43 0 .826-.216 1.078-.611l1.976-3.16c.359-.61.18-1.4-.396-1.76-.61-.394-1.4-.214-1.76.396l-1.976 3.196c-.395.574-.216 1.364.395 1.759zM5.29 7.289c.25.395.646.61 1.077.61.216 0 .467-.072.683-.18.61-.359.79-1.148.395-1.759L5.469 2.765c-.36-.61-1.15-.79-1.76-.395-.611.36-.791 1.15-.396 1.76l1.976 3.159z"/>
+    {/* <path fill="#F6F6F6" fill-rule="evenodd" d="M16.933 7.553c.17.114.342.146.572.146.36 0 .692-.178.909-.517l1.973-3.156c.307-.523.147-1.189-.33-1.487l-.002-.001c-.519-.335-1.18-.18-1.48.329l-.002.003-1.981 3.204c-.326.474-.185 1.139.339 1.478l.002.001zm3.625-3.424l-1.975 3.16c-.252.395-.647.61-1.078.61-.252 0-.467-.036-.683-.18-.61-.394-.79-1.184-.395-1.758l1.976-3.196c.36-.61 1.15-.79 1.76-.395.575.359.755 1.149.395 1.76zM7.275 6.065L5.296 2.866c-.3-.508-.96-.663-1.48-.328l-.006.005c-.51.299-.665.96-.33 1.478l.003.002 1.975 3.158c.216.34.549.518.909.518.175 0 .393-.06.587-.155.505-.301.657-.959.323-1.475l-.002-.004zM5.289 7.29c.252.395.647.61 1.078.61.216 0 .467-.072.683-.18.61-.359.79-1.148.395-1.759L5.469 2.765c-.36-.61-1.15-.79-1.76-.395-.611.36-.791 1.15-.396 1.76l1.976 3.159zm6.647-1.17c.608 0 1.093-.485 1.093-1.092V1.293C13.03.685 12.544.2 11.936.2s-1.093.485-1.093 1.093v3.77c0 .568.481 1.056 1.093 1.056zm0 .2c.719 0 1.293-.574 1.293-1.292V1.293C13.23.575 12.655 0 11.936 0s-1.293.575-1.293 1.293v3.77c0 .682.574 1.256 1.293 1.256z" clip-rule="evenodd"/> */}
   </svg>
 )
