+++ assets/javascripts/kitten/components/graphics/icons-next/gift-icon/index.js
