+++ assets/javascripts/kitten/components/graphics/icons-next/lock-outline-icon/index.js
