+++ assets/javascripts/kitten/components/graphics/icons-next/refund-icon/index.js
