+++ assets/javascripts/kitten/index.js
@@ -136,7 +136,6 @@ export { FilterIcon } from './components/graphics/icons/filter-icon'
 export { FlagIcon, flagList } from './components/graphics/icons/flag-icon'
 export { FlashCircleIcon } from './components/graphics/icons/flash-circle-icon'
 export { GarbageIcon } from './components/graphics/icons/garbage-icon'
-export { GiftIcon } from './components/graphics/icons/gift-icon'
 export { GiropayIcon } from './components/graphics/icons/giropay-icon'
 export { GlobeIcon } from './components/graphics/icons/globe-icon'
 export { GrabberIcon } from './components/graphics/icons/grabber-icon'
@@ -152,7 +151,6 @@ export { LinkIcon } from './components/graphics/icons/link-icon'
 export { ListIcon } from './components/graphics/icons/list-icon'
 export { LocationIcon } from './components/graphics/icons/location-icon'
 export { LockIcon } from './components/graphics/icons/lock-icon'
-export { LockOutlineIcon } from './components/graphics/icons/lock-outline-icon'
 export { LoginIcon } from './components/graphics/icons/login-icon'
 export { NoImageIcon } from './components/graphics/icons/no-image-icon'
 export { MaestroIcon } from './components/graphics/icons/maestro-icon'
@@ -164,7 +162,6 @@ export { PasswordIcon } from './components/graphics/icons/password-icon'
 export { PayPalIcon } from './components/graphics/icons/paypal-icon'
 export { PhoneIcon } from './components/graphics/icons/phone-icon'
 export { QuestionMarkIcon } from './components/graphics/icons/question-mark-icon'
-export { RefundIcon } from './components/graphics/icons/refund-icon'
 export { RocketCircleIcon } from './components/graphics/icons/rocket-circle-icon'
 export { RocketIcon } from './components/graphics/icons/rocket-icon'
 export { SaveIcon } from './components/graphics/icons/save-icon'
@@ -212,6 +209,7 @@ export { EnvelopeIconNext } from './components/graphics/icons-next/envelope-icon
 export { EyeIconNext } from './components/graphics/icons-next/eye-icon-next'
 export { FilterIconNext } from './components/graphics/icons-next/filter-icon-next'
 export { FacebookIconNext } from './components/graphics/icons-next/facebook-icon-next'
+export { GiftIcon } from './components/graphics/icons-next/gift-icon'
 export { GiftIconNext } from './components/graphics/icons-next/gift-icon-next'
 export { HeartIconNext } from './components/graphics/icons-next/heart-icon-next'
 export { HeartWithClickIconNext } from './components/graphics/icons-next/heart-icon-next/with-click'
@@ -220,8 +218,10 @@ export { InfiniteIconNext } from './components/graphics/icons-next/infinite-icon
 export { InstagramIconNext } from './components/graphics/icons-next/instagram-icon-next'
 export { LargeArrowIconNext } from './components/graphics/icons-next/large-arrow-icon-next'
 export { LinkedinIconNext } from './components/graphics/icons-next/linkedin-icon-next'
+export { LockOutlineIcon } from './components/graphics/icons-next/lock-outline-icon'
 export { MessengerIconNext } from './components/graphics/icons-next/messenger-icon-next'
 export { PlusIconNext } from './components/graphics/icons-next/plus-icon-next'
+export { RefundIcon } from './components/graphics/icons-next/refund-icon'
 export { SpeechBubbleIconNext } from './components/graphics/icons-next/speech-bubble-icon-next'
 export { StarIconNext } from './components/graphics/icons-next/star-icon-next'
 export { StatsIconNext } from './components/graphics/icons-next/stats-icon-next'
